requirements.txt
requests
fusepy
sanic-jinja2
